* Track visualizer start
* Add icon

+++ MDAWLib/System/Song.cs
@@ -33,7 +33,7 @@ namespace MDAWLib1
         }
 
         public virtual string Title { get; } = "untitled";
-        public virtual int? Duration { get; }
+        public virtual double? Duration { get; }
 
         private int sampleRate = DefaultSampleReate;
         private WaveFormat? waveFormat;
